Many changes - refactoring, bugfixing Arbitrary highlighting actually works now (on a limited test case)
Interface:
* more documentation
* removed the setStart* and setEnd* method from Range - use "start() =" and "end() ="
* notification of direct changes to cursors and ranges
* fix parent/child range system
* constrain / expand ranges as required, when one range is changed directly
* new notification contentsChanged() with more information - see docs
* added methods to work with QStringLists
* added method to get the QChar for a cursor
* added some access and editing methods to Smart* classes
* bugfixes
  * constrainToRange() was broken
  * others :)
* Renamed some methods:
  * Document:
    * length() -> totalCharacters()
    * all() -> documentRange()
    * end() -> documentEnd()

Kate part:
* port to new changes
* fix rendering of arbitrary highlighting (many bugs)

Kate test:
* new arbitrary highlighting test - insert some curly brackets to test :)

svn path=/trunk/KDE/kdelibs/kate/; revision=470815

+++ part/part/katetextlayout.cpp
@@ -183,7 +183,7 @@ int KateTextLayout::startX( ) const
   if (m_startX == -1)
     // viewLine is already > 0, from the constructor
     for (int i = 0; i < viewLine(); ++i)
-      m_startX += m_lineLayout->layout()->lineAt(i).naturalTextWidth();
+      m_startX += (int)m_lineLayout->layout()->lineAt(i).naturalTextWidth();
 
   return m_startX;
 }
@@ -193,7 +193,7 @@ int KateTextLayout::endX( ) const
   if (!isValid())
     return 0;
 
-  return m_startX + m_textLayout.naturalTextWidth();
+  return m_startX + (int)m_textLayout.naturalTextWidth();
 }
 
 int KateTextLayout::width( ) const
@@ -201,7 +201,7 @@ int KateTextLayout::width( ) const
   if (!isValid())
     return 0;
 
-  return m_textLayout.naturalTextWidth();
+  return (int)m_textLayout.naturalTextWidth();
 }
 
 KateTextLayout KateTextLayout::invalid( )
